JSON responses are now sane.
Adds a mixin that enables the JSON marshaller to unwrap the actual response
object out of the JAXBElement envelope prior to marshalling. Configures the
JSON marshaller to use the mixin and also to take advantage of the JAXB
annotation introspector which ensures uniformity of the response attributes
regardless of whether they are JSON or XML marshalled.